backend/requirements.txt
django
gunicorn
whitenoise[brotli]
djangorestframework
python-dotenv
psycopg